auth_api/python/requirements.txt
anyio==3.6.2
black==23.1.0
cffi==1.15.1
click==8.1.3
colorama==0.4.6
cryptography==39.0.0
ecdsa==0.18.0
fastapi==0.89.1
Flask==2.2.2
greenlet==2.0.2
h11==0.14.0
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mypy-extensions==0.4.3
mysqlclient==2.1.1
packaging==23.0
pathspec==0.11.0
platformdirs==2.6.2
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.4
python-dotenv==0.21.1
python-jose==3.3.0
python-multipart==0.0.5
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a32
sqlmodel==0.0.8
starlette==0.22.0
tomli==2.0.1
typing-extensions==4.4.0
uvicorn==0.20.0
Werkzeug==2.2.2
zipp==3.12.0
